requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
gunicorn==21.2.0
python-dotenv==1.0.1